Master

+++ src/main/java/com/jdg/mypage/repository/MainProjectSkillRepository.java
@@ -3,5 +3,5 @@ package com.jdg.mypage.repository;
 import com.jdg.mypage.entity.MainProjectSkillList;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MainProjectSkillRepository extends CrudRepository<MainProjectSkillList, Long> {
+public interface MainProjectSkillRepository extends CrudRepository<MainProjectSkillList, Integer> {
 }
